@@ -1,4 +1,5 @@
-import React, { Children, PropTypes } from 'react';
+import React, { Children } from 'react';
+import PropTypes from 'prop-types';
 import Theme, { ThemeShape } from './Theme';
 import normalizeStyle from './StyleNormalizer/normalizeStyle';
 
@@ -8,7 +9,7 @@ import normalizeStyle from './StyleNormalizer/normalizeStyle';
 export default class StyleProvider extends React.Component {
   static propTypes = {
     children: PropTypes.element.isRequired,
-    style: React.PropTypes.object,
+    style: PropTypes.object,
   };
 
   static defaultProps = {
