@@ -1,8 +1,8 @@
-import connectStyle from './src/connectStyle';
-import { INCLUDE } from './src/resolveIncludes';
-import StyleProvider from './src/StyleProvider';
-import Theme, { ThemeShape } from './src/Theme';
-import { createVariations, createSharedStyle } from './src/addons';
+import connectStyle from "./src/connectStyle";
+import { INCLUDE } from "./src/resolveIncludes";
+import StyleProvider from "./src/StyleProvider";
+import Theme, { ThemeShape } from "./src/Theme";
+import { createVariations, createSharedStyle } from "./src/addons";
 
 export {
   connectStyle,
