@@ -1,15 +1,16 @@
-import { PropTypes } from 'react';
-import resolveIncludes from './resolveIncludes';
-import mergeComponentAndThemeStyles from './mergeComponentAndThemeStyles';
-import normalizeStyle from './StyleNormalizer/normalizeStyle';
+import { PropTypes } from "prop-types";
+import resolveIncludes from "./resolveIncludes";
+import mergeComponentAndThemeStyles from "./mergeComponentAndThemeStyles";
+import normalizeStyle from "./StyleNormalizer/normalizeStyle";
 
 // Privates, ideally those should be symbols
-const THEME_STYLE = '@@shoutem.theme/themeStyle';
-const THEME_STYLE_CACHE = '@@shoutem.theme/themeCachedStyle';
+const THEME_STYLE = "@@shoutem.theme/themeStyle";
+const THEME_STYLE_CACHE = "@@shoutem.theme/themeCachedStyle";
 
 let defaultTheme;
 
-const resolveStyle = (style, baseStyle) => normalizeStyle(resolveIncludes(style, baseStyle));
+const resolveStyle = (style, baseStyle) =>
+  normalizeStyle(resolveIncludes(style, baseStyle));
 
 /**
  * The theme defines the application style, and provides methods to
@@ -76,7 +77,10 @@ export default class Theme {
       return this[THEME_STYLE_CACHE][componentName];
     }
 
-    const componentIncludedStyle = resolveStyle(defaultStyle, this[THEME_STYLE]);
+    const componentIncludedStyle = resolveStyle(
+      defaultStyle,
+      this[THEME_STYLE]
+    );
 
     /**
      * This is static component style (static per componentName). This style can only
@@ -94,5 +98,5 @@ export default class Theme {
 }
 
 export const ThemeShape = PropTypes.shape({
-  createComponentStyle: PropTypes.func.isRequired,
+  createComponentStyle: PropTypes.func.isRequired
 });
