@@ -1,10 +1,14 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import hoistStatics from 'hoist-non-react-statics';
 import * as _ from 'lodash';
 import normalizeStyle from './StyleNormalizer/normalizeStyle';
+import { StyleSheet } from "react-native";
 
-import Theme, { ThemeShape } from './Theme';
-import { resolveComponentStyle } from './resolveComponentStyle';
+import Theme, { ThemeShape } from "./Theme";
+import { resolveComponentStyle } from "./resolveComponentStyle";
+
+const themeCache = {};
 
 /**
  * Formats and throws an error when connecting component style with the theme.
@@ -13,7 +17,9 @@ import { resolveComponentStyle } from './resolveComponentStyle';
  * @param componentDisplayName The name of the component that is being connected.
  */
 function throwConnectStyleError(errorMessage, componentDisplayName) {
-  throw Error(`${errorMessage} - when connecting ${componentDisplayName} component to style.`);
+  throw Error(
+    `${errorMessage} - when connecting ${componentDisplayName} component to style.`
+  );
 }
 
 /**
@@ -27,6 +33,41 @@ function getTheme(context) {
   // Fallback to a default theme if the component isn't
   // rendered in a StyleProvider.
   return context.theme || Theme.getDefaultTheme();
+}
+
+/**
+ * Matches any style properties that represent component style variants.
+ * Those styles can be applied to the component by using the styleName
+ * prop. All style variant property names must start with a single '.'
+ * character, e.g., '.variant'.
+ *
+ * @param propertyName The style property name.
+ * @returns {boolean} True if the style property represents a component variant, false otherwise.
+ */
+function isStyleVariant(propertyName) {
+  return /^\./.test(propertyName);
+}
+
+/**
+ * Matches any style properties that represent style rules that target the
+ * component children. Those styles can have two formats, they can either
+ * target the components by component name ('shoutem.ui.Text'), or by component
+ * name and variant ('shoutem.ui.Text.line-through'). Beside specifying the
+ * component name, those styles can also target any component by using the
+ * '*' wildcard ('*', or '*.line-through'). The rule to identify those styles is
+ * that they have to contain a '.' character in their name or be a '*'.
+ *
+ * @param propertyName The style property name.
+ * @returns {boolean} True if the style property represents a child style, false otherwise.
+ */
+function isChildStyle(propertyName) {
+  return /(^[^\.].*\.)|^\*$/.test(propertyName);
+}
+
+function getConcreteStyle(style) {
+  return _.pickBy(style, (value, key) => {
+    return !isStyleVariant(key) && !isChildStyle(key);
+  });
 }
 
 /**
@@ -44,9 +85,14 @@ function getTheme(context) {
  * @returns {StyledComponent} The new component that will handle
  * the styling of the wrapped component.
  */
-export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, options = {}) => {
+export default (
+  componentStyleName,
+  componentStyle = {},
+  mapPropsToStyleNames,
+  options = {}
+) => {
   function getComponentDisplayName(WrappedComponent) {
-    return WrappedComponent.displayName || WrappedComponent.name || 'Component';
+    return WrappedComponent.displayName || WrappedComponent.name || "Component";
   }
 
   return function wrapWithStyledComponent(WrappedComponent) {
@@ -54,14 +100,14 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
 
     if (!_.isPlainObject(componentStyle)) {
       throwConnectStyleError(
-        'Component style must be plain object',
+        "Component style must be plain object",
         componentDisplayName
       );
     }
 
     if (!_.isString(componentStyleName)) {
       throwConnectStyleError(
-        'Component Style Name must be string',
+        "Component Style Name must be string",
         componentDisplayName
       );
     }
@@ -70,18 +116,24 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
       static contextTypes = {
         theme: ThemeShape,
         // The style inherited from the parent
-        parentStyle: PropTypes.object,
+        // parentStyle: PropTypes.object,
+        parentPath: PropTypes.array
       };
 
       static childContextTypes = {
         // Provide the parent style to child components
-        parentStyle: PropTypes.object,
-        resolveStyle: PropTypes.func,
+        // parentStyle: PropTypes.object,
+        // resolveStyle: PropTypes.func,
+        parentPath: PropTypes.array
       };
 
       static propTypes = {
         // Element style that overrides any other style of the component
-        style: PropTypes.object,
+        style: PropTypes.oneOfType([
+          PropTypes.object,
+          PropTypes.number,
+          PropTypes.array
+        ]),
         // The style variant names to apply to this component,
         // multiple variants may be separated with a space character
         styleName: PropTypes.string,
@@ -89,11 +141,11 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
         // style to their children, i.e., the children
         // will behave as they are placed directly below
         // the parent of a virtual element.
-        virtual: PropTypes.bool,
+        virtual: PropTypes.bool
       };
 
       static defaultProps = {
-        virtual: options.virtual,
+        virtual: options.virtual
       };
 
       static displayName = `Styled(${componentDisplayName})`;
@@ -101,38 +153,110 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
 
       constructor(props, context) {
         super(props, context);
-        const styleNames = this.resolveStyleNames(props);
-        const resolvedStyle = this.resolveStyle(context, props, styleNames);
+        // console.log(context.parentPath);
+        const styleNames = this.getStyleNames(props);
+        const style = props.style;
+
+        const finalStyle = this.getFinalStyle(
+          props,
+          context,
+          style,
+          styleNames
+        );
+
         this.setWrappedInstance = this.setWrappedInstance.bind(this);
-        this.resolveConnectedComponentStyle = this.resolveConnectedComponentStyle.bind(this);
+        this.resolveConnectedComponentStyle = this.resolveConnectedComponentStyle.bind(
+          this
+        );
         this.state = {
-          style: resolvedStyle.componentStyle,
-          childrenStyle: resolvedStyle.childrenStyle,
+          style: finalStyle,
           // AddedProps are additional WrappedComponent props
           // Usually they are set trough alternative ways,
           // such as theme style, or trough options
           addedProps: this.resolveAddedProps(),
-          styleNames,
+          styleNames
         };
+      }
+
+      getFinalStyle(props, context, style, styleNames) {
+        let resolvedStyle = {};
+        if (context.parentPath) {
+          resolvedStyle = this.getOrSetStylesInCache(
+            context,
+            props,
+            styleNames,
+            [...context.parentPath, componentStyleName, ...styleNames]
+          );
+        } else {
+          resolvedStyle = this.resolveStyle(context, props, styleNames);
+          themeCache[componentStyleName] = resolvedStyle;
+        }
+
+        const concreteStyle = getConcreteStyle(_.merge({}, resolvedStyle));
+
+        if (_.isArray(style)) {
+          return [concreteStyle, ...style];
+        }
+
+        if (typeof style == "number" || typeof style == "object") {
+          return [concreteStyle, style];
+        }
+
+        return concreteStyle;
+      }
+
+      getStyleNames(props) {
+        const styleNamesArr = _.map(props, (value, key) => {
+          if (typeof value !== "object" && value === true) {
+            return "." + key;
+          } else {
+            return false;
+          }
+        });
+        _.remove(styleNamesArr, (value, index) => {
+          return value === false;
+        });
+
+        return styleNamesArr;
+      }
+
+      getParentPath() {
+        if (!this.context.parentPath) {
+          return [componentStyleName];
+        } else {
+          return [
+            ...this.context.parentPath,
+            componentStyleName,
+            ...this.getStyleNames(this.props)
+          ];
+        }
       }
 
       getChildContext() {
         return {
-          parentStyle: this.props.virtual ?
-            this.context.parentStyle :
-            this.state.childrenStyle,
-          resolveStyle: this.resolveConnectedComponentStyle,
+          // parentStyle: this.props.virtual ?
+          //   this.context.parentStyle :
+          //   this.state.childrenStyle,
+          // resolveStyle: this.resolveConnectedComponentStyle,
+          parentPath: this.getParentPath()
         };
       }
 
       componentWillReceiveProps(nextProps, nextContext) {
-        const styleNames = this.resolveStyleNames(nextProps);
+        const styleNames = this.getStyleNames(nextProps);
+        const style = nextProps.style;
         if (this.shouldRebuildStyle(nextProps, nextContext, styleNames)) {
-          const resolvedStyle = this.resolveStyle(nextContext, nextProps, styleNames);
+          const finalStyle = this.getFinalStyle(
+            nextProps,
+            nextContext,
+            style,
+            styleNames
+          );
+
           this.setState({
-            style: resolvedStyle.componentStyle,
-            childrenStyle: resolvedStyle.childrenStyle,
-            styleNames,
+            style: finalStyle,
+            // childrenStyle: resolvedStyle.childrenStyle,
+            styleNames
           });
         }
       }
@@ -144,26 +268,32 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
       }
 
       setWrappedInstance(component) {
-        if(component && component._root) {
+        if (component && component._root) {
           this._root = component._root;
         } else {
           this._root = component;
         }
+        this.wrappedInstance = this._root;
       }
 
       hasStyleNameChanged(nextProps, styleNames) {
-        return mapPropsToStyleNames && this.props !== nextProps &&
+        return (
+          mapPropsToStyleNames &&
+          this.props !== nextProps &&
           // Even though props did change here,
           // it doesn't necessary means changed props are those which affect styleName
-          !_.isEqual(this.state.styleNames, styleNames);
+          !_.isEqual(this.state.styleNames, styleNames)
+        );
       }
 
       shouldRebuildStyle(nextProps, nextContext, styleNames) {
-        return (nextProps.style !== this.props.style) ||
-          (nextProps.styleName !== this.props.styleName) ||
-          (nextContext.theme !== this.context.theme) ||
-          (nextContext.parentStyle !== this.context.parentStyle) ||
-          (this.hasStyleNameChanged(nextProps, styleNames));
+        return (
+          nextProps.style !== this.props.style ||
+          nextProps.styleName !== this.props.styleName ||
+          nextContext.theme !== this.context.theme ||
+          !_.isEqual(nextContext.parentPath, this.context.parentPath) ||
+          this.hasStyleNameChanged(nextProps, styleNames)
+        );
       }
 
       resolveStyleNames(props) {
@@ -181,24 +311,50 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
       resolveAddedProps() {
         const addedProps = {};
         if (options.withRef) {
-          addedProps.ref = 'wrappedInstance';
+          addedProps.ref = "wrappedInstance";
         }
         return addedProps;
       }
 
+      getOrSetStylesInCache(context, props, styleNames, path) {
+        if (themeCache && themeCache[path.join(">")]) {
+          // console.log('**************');
+
+          return themeCache[path.join(">")];
+        } else {
+          resolvedStyle = this.resolveStyle(context, props, styleNames);
+          if (Object.keys(themeCache).length < 10000) {
+            themeCache[path.join(">")] = resolvedStyle;
+          }
+          return resolvedStyle;
+        }
+      }
+
       resolveStyle(context, props, styleNames) {
-        const { parentStyle } = context;
-        const style = normalizeStyle(props.style);
+        let parentStyle = {};
 
         const theme = getTheme(context);
-        const themeStyle = theme.createComponentStyle(componentStyleName, componentStyle);
+        const themeStyle = theme.createComponentStyle(
+          componentStyleName,
+          componentStyle
+        );
+
+        if (context.parentPath) {
+          parentStyle = themeCache[context.parentPath.join(">")];
+        } else {
+          parentStyle = resolveComponentStyle(
+            componentStyleName,
+            styleNames,
+            themeStyle,
+            parentStyle
+          );
+        }
 
         return resolveComponentStyle(
           componentStyleName,
           styleNames,
           themeStyle,
-          parentStyle,
-          style
+          parentStyle
         );
       }
 
@@ -211,10 +367,18 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
        */
       resolveConnectedComponentStyle(props) {
         const styleNames = this.resolveStyleNames(props);
-        return this.resolveStyle(this.context, props, styleNames).componentStyle;
+        return this.resolveStyle(this.context, props, styleNames)
+          .componentStyle;
       }
 
       render() {
+        // console.log('themeCache', themeCache);
+
+        // if(componentStyleName == 'NativeBase.Text') {
+        //   console.log(this.state.style);
+        //   console.log(themeCache);
+        // }
+
         const { addedProps, style } = this.state;
         return (
           <WrappedComponent
@@ -222,7 +386,8 @@ export default (componentStyleName, componentStyle = {}, mapPropsToStyleNames, o
             {...addedProps}
             style={style}
             ref={this.setWrappedInstance}
-          />);
+          />
+        );
       }
     }
 
