@@ -1,4 +1,5 @@
-import _ from 'lodash';
+import _ from "lodash";
+import customMerge from "./customMerge";
 
 /**
  * Matches any style properties that represent component style variants.
@@ -39,20 +40,24 @@ function isChildStyle(propertyName) {
  * @returns {*} An object with the componentStyle, styleVariants, and childrenStyle keys.
  */
 function splitStyle(style) {
-  return _.reduce(style, (result, value, key) => {
-    let styleSection = result.componentStyle;
-    if (isStyleVariant(key)) {
-      styleSection = result.styleVariants;
-    } else if (isChildStyle(key)) {
-      styleSection = result.childrenStyle;
+  return _.reduce(
+    style,
+    (result, value, key) => {
+      let styleSection = result.componentStyle;
+      if (isStyleVariant(key)) {
+        styleSection = result.styleVariants;
+      } else if (isChildStyle(key)) {
+        styleSection = result.childrenStyle;
+      }
+      styleSection[key] = value;
+      return result;
+    },
+    {
+      componentStyle: {},
+      styleVariants: {},
+      childrenStyle: {}
     }
-    styleSection[key] = value;
-    return result;
-  }, {
-    componentStyle: {},
-    styleVariants: {},
-    childrenStyle: {},
-  });
+  );
 }
 
 /**
@@ -83,40 +88,53 @@ export function resolveComponentStyle(
   styleNames = [],
   themeStyle = {},
   parentStyle = {},
-  elementStyle = {}
+  themeCache
 ) {
-  // Phase 1: merge the styles in the correct order to resolve the variant styles,
-  // the component style will be merged as well in this step, but the component
-  // style merge results are ignored after this step. We need to perform this
-  // step separately because the style variants may be overridden by any style, so
-  // the purpose of this phase is to determine the final state of the variant styles.
-  const mergedStyle = _.merge({},
-    themeStyle,
-    parentStyle['*'],
-    parentStyle[componentName],
-    ..._.map(styleNames, (sn) => themeStyle[`.${sn}`]),
-    ..._.map(styleNames, (sn) => parentStyle[`*.${sn}`]),
-    ..._.map(styleNames, (sn) => parentStyle[`${componentName}.${sn}`]),
-    elementStyle
-  );
+  // const mergedStyle = _.merge({},
+  //   themeStyle,
+  //   parentStyle['*'],
+  //   parentStyle[componentName],
+  //   ..._.map(styleNames, (sn) => themeStyle[`.${sn}`]),
+  //   ..._.map(styleNames, (sn) => parentStyle[`*.${sn}`]),
+  //   ..._.map(styleNames, (sn) => parentStyle[`${componentName}.${sn}`])
+  // );
+
+  let mergedStyle = customMerge(themeStyle, parentStyle[componentName]);
+  styleNames.forEach((sn, index) => {
+    mergedStyle = customMerge(mergedStyle, themeStyle[`${sn}`]);
+  });
+
+  styleNames.forEach((sn, index) => {
+    mergedStyle = customMerge(
+      mergedStyle,
+      parentStyle[`${componentName}${sn}`]
+    );
+  });
 
   // Phase 2: merge the component styles, this step is performed by using the
   // style from phase 1, so that we are sure that the final style variants are
   // applied to component style.
-  const resolvedStyle = _.merge({},
-    mergedStyle,
-    parentStyle['*'],
-    parentStyle[componentName],
-    ..._.map(styleNames, (sn) => mergedStyle[`.${sn}`]),
-    ..._.map(styleNames, (sn) => parentStyle[`*.${sn}`]),
-    ..._.map(styleNames, (sn) => parentStyle[`${componentName}.${sn}`]),
-    elementStyle
-  );
+  // const resolvedStyle = _.merge({},
+  //   mergedStyle,
+  //   parentStyle['*'],
+  //   parentStyle[componentName],
+  //   ..._.map(styleNames, (sn) => mergedStyle[`.${sn}`]),
+  //   ..._.map(styleNames, (sn) => parentStyle[`*.${sn}`]),
+  //   ..._.map(styleNames, (sn) => parentStyle[`${componentName}.${sn}`])
+  // );
 
-  const { componentStyle, childrenStyle } = splitStyle(resolvedStyle);
+  let resolvedStyle = customMerge(mergedStyle, parentStyle[componentName]);
 
-  return {
-    componentStyle,
-    childrenStyle,
-  };
+  styleNames.forEach((sn, index) => {
+    resolvedStyle = customMerge(resolvedStyle, mergedStyle[`${sn}`]);
+  });
+
+  styleNames.forEach((sn, index) => {
+    resolvedStyle = customMerge(
+      resolvedStyle,
+      parentStyle[`${componentName}${sn}`]
+    );
+  });
+
+  return resolvedStyle;
 }
